Fix eresource, checkin record numbers
Running the entire test suite, I found the changes we made to record
numbers during the refactor (i.e., they no longer employ check digits)
messed up a handful of tests. I've updating things so everything is
passing now.

+++ django/sierra/base/models.py
@@ -508,7 +508,7 @@ class RecordMetadata(ReadOnlyModel):
         rn = int(self.record_num)
         full_rec_num = '{}{}'.format(self.record_type, self.record_num)
         if use_check_digit:
-            check_digit = (sum([(rn / (10 ** i) % 10) * (i + 2)
+            check_digit = int(sum([(rn / (10 ** i) % 10) * (i + 2)
                                 for i in range(0, 6)])
                            + (rn / (10 ** 6)) * 8) % 11
             if check_digit == 10:
